refactor(safety_checker): remove lane change dependency (#4387)
* refactor(safety_checker): remove lane change dependency


* refactor(safety_checker): init primitive member variables


---------
